refactor: Create shinyBindings.jsx for the relevant code
