+++ js/src/react/index.js
@@ -1,3 +1,3 @@
 export * from './adapters';
 export { default as findAndRenderReactData } from './findAndRenderReactData';
-export { default as ShinyBindingWrapper } from './ShinyBindingWrapper';
+export { ShinyBindingWrapper } from './shinyBindings';
